pyproject.toml
[project]
name = "smart-web-attack-replay-generator"
version = "1.1.0"
description = "Smart Web Application Attack Replay Generator - Parse logs, detect attacks, generate replay scripts with ML and real-time monitoring"
requires-python = ">=3.11"
dependencies = [
    # Core web framework
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Machine Learning
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    
    # Real-time monitoring
    "watchdog>=3.0.0",
    
    # HTTP requests
    "requests>=2.31.0",
    
    # Database
    "psycopg2-binary>=2.9.11",
    
    # WebSocket support
    "python-socketio>=5.8.0",
    "eventlet>=0.33.0",
]
